JENKINS-44419 - Organize clases into correct packages
